backend/backend/requirements.txt
keyring
djangorestframework-simplejwt
psutil
pySMART
prometheus_client
paramiko
fabric
Django
djangorestframework
psycopg2-binary
redis
celery
django-celery-beat
django-celery-results
channels
daphne
django-oauth-toolkit
PyJWT
gunicorn
drf-yasg
django-redis
python-decouple
django-cors-headers
django-environ
speedtest-cli
